requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
Faker==37.6.0
Flask==3.1.2
groq==0.31.1
h11==0.16.0
httpagentparser==1.9.5
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==2.4.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.6
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.1
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
gensim